requirements.txt
flask
pillow
pillow-heif
pypdf
qrcode[pil]
gunicorn
mysql-connector-python
python-dotenv